fix wrong unsigned int size

+++ packager/media/formats/mp4/segmenter.h
@@ -144,8 +144,8 @@ class Segmenter {
   ProgressListener* progress_listener_ = nullptr;
   uint64_t progress_target_ = 0u;
   uint64_t accumulated_progress_ = 0u;
-  uint32_t sample_durations_[2] = {0, 0};
-  size_t num_samples_ = 0u;
+  int64_t sample_durations_[2] = {0, 0};
+  size_t num_samples_ = 0;
   std::vector<uint64_t> stream_durations_;
   std::vector<KeyFrameInfo> key_frame_infos_;
